+++ rest/core/src/main/java/org/seedstack/seed/rest/internal/RestModule.java
@@ -11,22 +11,18 @@ package org.seedstack.seed.rest.internal;
 import com.google.inject.AbstractModule;
 import com.google.inject.Scopes;
 import com.google.inject.multibindings.MapBinder;
-import com.google.inject.name.Names;
 import java.util.Collection;
 import java.util.Map;
 import javax.ws.rs.core.Variant;
-import org.seedstack.seed.rest.RestConfig;
 import org.seedstack.seed.rest.spi.RootResource;
 
 class RestModule extends AbstractModule {
     private final Collection<Class<?>> resources;
     private final Collection<Class<?>> providers;
-    private final RestConfig restConfig;
     private final Map<Variant, Class<? extends RootResource>> rootResourcesByVariant;
 
-    RestModule(RestConfig restConfig, Collection<Class<?>> resources, Collection<Class<?>> providers,
+    RestModule(Collection<Class<?>> resources, Collection<Class<?>> providers,
             Map<Variant, Class<? extends RootResource>> rootResourcesByVariant) {
-        this.restConfig = restConfig;
         this.rootResourcesByVariant = rootResourcesByVariant;
         this.resources = resources;
         this.providers = providers;
@@ -34,17 +30,12 @@ class RestModule extends AbstractModule {
 
     @Override
     protected void configure() {
-        bind(String.class).annotatedWith(Names.named("SeedRestPath")).toInstance(restConfig.getPath());
-        bind(String.class).annotatedWith(Names.named("SeedJspPath")).toInstance(restConfig.getJspPath());
-
         for (Class<?> resource : resources) {
             bind(resource);
         }
-
         for (Class<?> provider : providers) {
             bind(provider).in(Scopes.SINGLETON);
         }
-
         if (!rootResourcesByVariant.isEmpty()) {
             MapBinder<Variant, RootResource> multiBinder = MapBinder.newMapBinder(binder(), Variant.class,
                     RootResource.class);