+++ rest/jersey2/src/test/java/org/seedstack/seed/rest/jersey2/CorsRestIT.java
@@ -20,7 +20,7 @@ import org.seedstack.seed.undertow.LaunchWithUndertow;
 @LaunchWithUndertow
 public class CorsRestIT {
     private static final String WWW_OTHER_COM = "www.other.com";
-    @Configuration("web.runtime.baseUrl")
+    @Configuration("runtime.web.baseUrl")
     private String baseUrl;
 
     @Test