+++ rest/jersey2/src/test/java/org/seedstack/seed/rest/jersey2/ErroneousResourceIT.java
@@ -19,7 +19,7 @@ import org.seedstack.seed.undertow.LaunchWithUndertow;
 @RunWith(SeedITRunner.class)
 @LaunchWithUndertow
 public class ErroneousResourceIT {
-    @Configuration("web.runtime.baseUrl")
+    @Configuration("runtime.web.baseUrl")
     private String baseUrl;
 
     @Test