+++ rest/jersey2/src/test/java/org/seedstack/seed/rest/jersey2/Jersey2IT.java
@@ -24,7 +24,7 @@ import org.skyscreamer.jsonassert.JSONAssert;
 @RunWith(SeedITRunner.class)
 @LaunchWithUndertow
 public class Jersey2IT {
-    @Configuration("web.runtime.baseUrl")
+    @Configuration("runtime.web.baseUrl")
     private String baseUrl;
 
     @Test