+++ rest/jersey2/src/test/java/org/seedstack/seed/rest/jersey2/WithoutRootResourceIT.java
@@ -21,7 +21,7 @@ import org.seedstack.seed.undertow.LaunchWithUndertow;
 @LaunchWithUndertow
 @SystemProperty(name = "seedstack.profiles", value = "withoutRootResource")
 public class WithoutRootResourceIT {
-    @Configuration("web.runtime.baseUrl")
+    @Configuration("runtime.web.baseUrl")
     private String baseUrl;
 
     @Test