+++ web/core/src/main/java/org/seedstack/seed/web/internal/WebRuntimeConfigurationProvider.java
@@ -40,8 +40,10 @@ class WebRuntimeConfigurationProvider implements ConfigurationProvider {
             }
         }
 
-        return new MapNode(new NamedNode("web", new MapNode(
-                new NamedNode("runtime", info)
+        return new MapNode(new NamedNode("runtime", new MapNode(
+                new NamedNode("web", new MapNode(
+                        new NamedNode("servlet", info)
+                ))
         )));
     }
 }